pyproject.toml
[tool.poetry]
name = "infx-semantic-normalization-api"
version = "0.0.1"
description = "INFX Semantic Normalization Service"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.7"
flask = "^3.0.1"
waitress = "^2.1.2"
python-decouple = "^3.8"


[[tool.poetry.source]]
name = "internal_pypi"
url = "https://repo.devops.projectronin.io/repository/pypi/simple/"
priority = "explicit"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
coverage = "^7.4.3"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
